attribute comparison returns Double

+++ src/main/scala-2.11/relationalClustering/similarity/SimilarityNeighbourhoodTrees.scala
@@ -78,7 +78,7 @@ class SimilarityNeighbourhoodTrees(override protected val knowledgeBase: Knowled
     val attrs1 = ng1.getRootAttributes
     val attrs2 = ng2.getRootAttributes
 
-    attrs1.intersect(attrs2).size
+    attrs1.intersect(attrs2).size.toDouble
   }
 
   /** Computes the attribute neighbourhood similarity of two neighbourhood graphs, per level and vertex type
